% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misreport_summary.R
\name{misreport_summary}
\alias{misreport_summary}
\title{Summarize misreporting percentages}
\usage{
misreport_summary(data, group = NULL)
}
\arguments{
\item{data}{A data frame containing logical columns Underreporting, Overreporting, Misreporting.}

\item{group}{Optional grouping column (unquoted).}
}
\value{
A tibble with percentages and Total.
}
\description{
Calculates \% underreporting, \% overreporting, \% misreporting, and total N,
optionally by a grouping variable.
}
